[#65] Refactor: 다이어리 댓글 작성 컴포넌트 분리, 공용화
