+++ components/diaries/diary-content/comment/diary-comment-form.module.scss
@@ -1,0 +1,23 @@
+.commentForm {
+  position: relative;
+}
+
+.commentInput {
+  width: 100%;
+  height: 44px;
+  border: 1px solid var(--text-color-gray);
+  border-radius: 9999px;
+  padding: 0 24px;
+
+  &:focus-visible {
+    outline: none;
+    border: 1px solid var(--main-color);
+  }
+}
+
+.commentButton {
+  position: absolute;
+  right: 24px;
+  top: 50%;
+  transform: translateY(-45%);
+}
