+++ pages/diaries/[diaryId]/diary-id.module.scss
@@ -33,27 +33,3 @@
   min-height: 100px;
   word-break: break-all;
 }
-
-.commentForm {
-  position: relative;
-}
-
-.commentInput {
-  width: 100%;
-  height: 44px;
-  border: 1px solid var(--text-color-gray);
-  border-radius: 9999px;
-  padding: 0 24px;
-
-  &:focus-visible {
-    outline: none;
-    border: 1px solid var(--main-color);
-  }
-}
-
-.commentButton {
-  position: absolute;
-  right: 24px;
-  top: 50%;
-  transform: translateY(-45%);
-}
